display methods for product index delocalized in helper modules
